Move security manager override to its own module (#32987)
(cherry picked from commit c10f67b36ad101a5f447a5dc075a02fffec7d0f0)
