+++ airflow/auth/managers/fab/models/anonymous_user.py
@@ -20,7 +20,7 @@ from __future__ import annotations
 from flask import current_app
 from flask_login import AnonymousUserMixin
 
-from airflow.auth.managers.model.base_user import BaseUser
+from airflow.auth.managers.models.base_user import BaseUser
 
 
 class AnonymousUser(AnonymousUserMixin, BaseUser):
